third_party/WebKit/LayoutTests/fast/dom/custom/invalid-first-char-combinators.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>

<script>
test(function () {
    var invalidCombinators = [ 0x0B83, 0x0F88, 0x0F89, 0x0F8A, 0x0F8B ];

    for (var i = 0; i < invalidCombinators.length; i++) {
        var invalidCombinator = String.fromCharCode(invalidCombinators[i]) + '-xfoo';
        assert_throws('SYNTAX_ERR', function () {
            document.registerElement(invalidCombinator);
        }, 'registering custom element: *' +
        invalidCombinator + '* should throw syntax error');
    }

}, 'registering invalid first letter combinators, not covered in Document::validNames');
</script>
